Move kernel_matrix and square_kernel_matrix
The methods now live in a separate utils file.
